move all the library files
