+++ cnc_ctrl_v1/CNC_Functions.h
@@ -16,9 +16,9 @@
 libraries*/
     
     
-#include "GearMotor.h"
-#include "Axis.h"
-#include "Kinematics.h"
+#include "libraries/GearMotor/GearMotor.h"
+#include "libraries/Axis/Axis.h"
+#include "libraries/Kinematics/Kinematics.h"
 
 #define ZAXIS
