+++ cnc_ctrl_v1/libraries/Axis/axis.h
@@ -19,10 +19,10 @@
     #define Axis_h
 
     #include "Arduino.h"
-    #include "GearMotor.h"
-    #include "PID_v1.h"
+    #include "../GearMotor/GearMotor.h"
+    #include "../PID/PID_v1.h"
     #include <EEPROM.h>
-    #include "Encoder.h"
+    #include "../Encoder/Encoder.h"
     
 
     class Axis{
