+++ cnc_ctrl_v1/libraries/Encoder/Encoder.cpp
